Migrate contents from dev.launchpad.net
Of the more than 1,000 pages, only the most important ones were migrated.

There will be an HTML dump of the old site, so we are able to look up
things, we might have missed to migrate.

Also, the migrated pages are not yet fully revised, so need some further
work, but this should be a good start.