edit: Skipping linkchecker CI on pull request. (#257)
This change is necessary because Launchpad has many links that, for multiple reasons, timeout during CI tests. This leads to long delays even for pull requests where changes don't include any links. As a result, this pull request will disable running of the link check on pull requests and replace it with a linkcheck via GitHub action that runs daily at 2200 UTC. This will ensure we stay updated on the state of links in the repository without blocking contributions.