Add documentation on how to build snaps in Launchpad (#262)
* Add documentation on how to build snaps in Launchpad

This document aims to provide a clear guidance on why snaps should be
built on Launchpad, the pre-requisites needed to do so, and the steps
to build them via the API and the UI.

* Update the create snaps using API and fix other typos and mistakes

Update the commands to create snaps using the API with two methods.
Add words to the custom wordlist.
Fix spelling mistakes.
Change REPOSITORY_NAME to repository_name and PROJECT_NAME to
project_name to match the lowercase username.

* edit: Implementing documentation review changes.

Signed-off-by: odadacharles <charles.odada@canonical.com>

* Fix syntax issues and typing mistakes

* chore: Fixing typos and adding Snapcraft to wordlist.

Signed-off-by: odadacharles <charles.odada@canonical.com>

* Apply suggestions from code review

Removing references to Ubuntu Pro

* edit: Moved snaps document to user how-to section.

Signed-off-by: odadacharles <charles.odada@canonical.com>

---------

Signed-off-by: odadacharles <charles.odada@canonical.com>
Co-authored-by: odadacharles <charles.odada@canonical.com>